requirements.txt
openai==0.27.2
streamlit==1.20.0
streamlit-audiorecorder==0.0.2
